[Fix] change print format

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -12,7 +12,7 @@ size_t print_dlistint(const dlistint_t *h)
 		return (count);
 	while (h)
 	{
-		printf("%u\n", (unsigned int) h->n);
+		printf("%d\n",h->n);
 		h = h->next;
 		count++;
 	}
